pyproject.toml
[tool.poetry]
name = "discord-swear-detection-bot"
version = "0.2.0"
description = "Discord bot detecting swears & typos in chat messages"
authors = ["Lilith Roth <lilith@roth.systems>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.2"
"discord.py" = "^2.4.0"
django-environ = "^0.11.2"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pylint = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
